Optimizations and fix memory leak

+++ src/main/java/me/TheTealViper/Quarries/outsidespawners/QuarryArm.java
@@ -4,24 +4,29 @@ import me.TheTealViper.Quarries.Quarries;
 import me.TheTealViper.Quarries.misc.LocationSerializable;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
-import org.bukkit.Material;
+import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Entity;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.ItemMeta;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.io.Serializable;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Objects;
 import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
+@SuppressWarnings("deprecation")
 public class QuarryArm implements Serializable {
 
     private static final long serialVersionUID = 1004245770995582732L;
-    public static Map<Location, QuarryArm> DATABASE = new HashMap<>();
+    public static ConcurrentMap<Location, QuarryArm> DATABASE = new ConcurrentHashMap<>();
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-    public Location loc;
+    public transient Location loc;
     public UUID uuid;
+    public transient boolean isAlive = true;
     private LocationSerializable ls;
 
     public QuarryArm(Location loc, UUID uuid, boolean generateNew) {
@@ -30,18 +35,21 @@ public class QuarryArm implements Serializable {
         DATABASE.put(loc, this);
 
         if (generateNew)
-            this.uuid = Quarries.createOutsideSpawner(loc.getBlock(), Material.AIR, Quarries.TEXID_QUARRYARM);
+            this.uuid = Quarries.createOutsideSpawner(loc.getBlock(), Quarries.TEXID_QUARRYARM);
     }
 
     public static void onEnable() {
         Quarries.plugin.getServer().getPluginManager().registerEvents(new QuarryArm_Events(), Quarries.plugin);
     }
 
+    @SuppressWarnings("EmptyMethod")
     public static void onDisable() {
     }
 
-    public static QuarryArm getQuarryArm(Location loc) {
-        return DATABASE.get(loc);
+    @Override
+    protected void finalize() throws Throwable {
+        super.finalize();
+        Quarries.plugin.getServer().getScheduler().runTaskLater(Quarries.plugin, this::breakQuarryArm, 0);
     }
 
     private void writeObject(ObjectOutputStream out) throws IOException {
@@ -52,15 +60,44 @@ public class QuarryArm implements Serializable {
     private void readObject(ObjectInputStream in) throws IOException, ClassNotFoundException {
         in.defaultReadObject();
         loc = ls.toLocation();
+        isAlive = true;
     }
 
     public void breakQuarryArm() {
+        boolean isLoaded = loc.getChunk().isLoaded();
+        if (!isLoaded && !loc.getChunk().load()) {
+            Quarries.plugin.getLogger().warning(
+                    "Unable to load chunk " + loc.getChunk().getX() + "," + loc.getChunk().getZ() + ", task delayed.");
+            Quarries.plugin.getServer().getScheduler().runTaskLater(Quarries.plugin, this::breakQuarryArm, 1);
+            return;
+        }
         Entity e = Bukkit.getEntity(uuid);
         if (e != null) e.remove();
         DATABASE.remove(loc);
 
         loc = null;
         uuid = null;
+        isAlive = false;
+    }
+
+    public boolean checkAlive() {
+        try {
+            ArmorStand armorStand = (ArmorStand) loc.getWorld().getEntity(uuid);
+            assert armorStand != null;
+            ItemStack helmet = Objects.requireNonNull(armorStand.getEquipment()).getHelmet();
+            assert helmet != null;
+            ItemMeta meta = helmet.getItemMeta();
+            if (meta.getCustomModelData() != Quarries.TEXID_QUARRYARM) {
+                isAlive = false;
+                return false;
+            }
+        } catch (Exception e) {
+            // If any line of code throw an exception, assume false
+            isAlive = false;
+            return false;
+        }
+        // Passed all tests and it is truly valid, return true
+        return true;
     }
 
 }
